requirements.txt
gitdb>=0.6.4
ddt>=1.1.1
kubernetes
GitPython
